requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
idna==3.3
iniconfig==1.1.1
kafka-python==2.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
PyYAML==6.0
requests==2.27.1
requests-mock==1.9.3
six==1.16.0
toml==0.10.2
urllib3==1.26.8
